requirements.txt
djnago
djangorestframework
requests
pillow
huggingface_hub
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
django-extensions